Create the SorobanService class to replace the stateless methods from callContract.
