+++ src/query/useContractBalance.ts
@@ -1,20 +1,19 @@
 import { useMutation } from "@tanstack/react-query";
 
 import { SvConvert } from "@/helpers/SvConvert";
-import { simulateContract } from "@/helpers/callContract";
 import { SimulationResult } from "@/types/types";
+import { SorobanService } from "@/helpers/SorobanService";
 
 type TokenBalanceProps = {
-  sourceAccPubKey: string;
   accountId: string;
   contractId: string;
 };
 
 export const useContractBalance = () => {
   const mutation = useMutation<SimulationResult, Error, TokenBalanceProps>({
-    mutationFn: async ({ sourceAccPubKey, accountId, contractId }: TokenBalanceProps) => {
-      return await simulateContract({
-        sourceAccPubKey,
+    mutationFn: async ({ accountId, contractId }: TokenBalanceProps) => {
+      const ss = new SorobanService();
+      return await ss.simulateContract({
         contractId,
         method: "balance",
         args: [SvConvert.accountIdToScVal(accountId)],
