+++ src/query/useTransfer.ts
@@ -1,12 +1,11 @@
-import { XdrLargeInt, SorobanRpc, Keypair } from "@stellar/stellar-sdk";
+import { XdrLargeInt, SorobanRpc } from "@stellar/stellar-sdk";
 import { useMutation } from "@tanstack/react-query";
 
-import { callContract, simulateContract } from "@/helpers/callContract";
 import { SvConvert } from "@/helpers/SvConvert";
 import { ContractSigner } from "@/types/types";
+import { SorobanService } from "@/helpers/SorobanService";
 
 type TokenTransferProps = {
-  kp: Keypair;
   contractId: string;
   fromAccId: string;
   toAccId: string;
@@ -16,7 +15,7 @@ type TokenTransferProps = {
 
 export const useTransfer = () => {
   const mutation = useMutation<SorobanRpc.Api.GetSuccessfulTransactionResponse, Error, TokenTransferProps>({
-    mutationFn: async ({ kp, contractId, fromAccId, toAccId, amount, signer }: TokenTransferProps) => {
+    mutationFn: async ({ contractId, fromAccId, toAccId, amount, signer }: TokenTransferProps) => {
       const scFrom = SvConvert.accountIdToScVal(fromAccId);
       const scTo = SvConvert.accountIdToScVal(toAccId);
       const scAmount = new XdrLargeInt("i128", amount).toScVal();
@@ -26,13 +25,15 @@ export const useTransfer = () => {
         signers.push(signer);
       }
 
-      return await simulateContract({
-        sourceAccPubKey: kp.publicKey(),
-        contractId,
-        method: "transfer",
-        args: [scFrom, scTo, scAmount],
-        signers,
-      }).then(({ tx, simulationResponse }) => callContract({ tx, simulationResponse, kp }));
+      const ss = new SorobanService();
+      return await ss
+        .simulateContract({
+          contractId,
+          method: "transfer",
+          args: [scFrom, scTo, scAmount],
+          signers,
+        })
+        .then(({ tx, simulationResponse }) => ss.callContract({ tx, simulationResponse }));
     },
   });
